+++ src/main/java/de/alpharogroup/compare/object/CompareObjectExtensions.java
@@ -1,17 +1,17 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
  * including without limitation the rights to use, copy, modify, merge, publish, distribute,
  * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all copies or
  * substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
  * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
  * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
@@ -21,20 +21,21 @@
 package de.alpharogroup.compare.object;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.Map;
+import java.util.Set;
 
 import org.apache.commons.beanutils.BeanComparator;
 import org.apache.commons.beanutils.BeanUtils;
 
 import de.alpharogroup.comparators.ComparatorExtensions;
-import lombok.experimental.UtilityClass;
 
 /**
  * The class {@link CompareObjectExtensions} provide methods for compare an object with another
  * given object.
  */
-@UtilityClass
 public final class CompareObjectExtensions
 {
 
@@ -79,7 +80,33 @@ public final class CompareObjectExtensions
 	}
 
 	/**
-	 * Compares the given two objects and returns the result as {@link int}.
+	 * Compares the given object over the given property.
+	 *
+	 * @param sourceOjbect
+	 *            the source ojbect
+	 * @param objectToCompare
+	 *            the object to compare
+	 * @param properties
+	 *            properties to compare
+	 * @return the resulted int value
+	 * @throws IllegalAccessException
+	 *             Thrown if this {@code Method} object is enforcing Java language access control
+	 *             and the underlying method is inaccessible.
+	 * @throws InvocationTargetException
+	 *             Thrown if the property accessor method throws an exception
+	 * @throws NoSuchMethodException
+	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
+	 *             "&lt;clinit&gt;".
+	 */
+	public static int compare(final Object sourceOjbect, final Object objectToCompare,
+		final String... properties)
+		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
+	{
+		return compareTo(sourceOjbect, objectToCompare, new HashSet<>(Arrays.asList(properties)));
+	}
+
+	/**
+	 * Compares the given two objects and returns the result as int.
 	 *
 	 * @param sourceOjbect
 	 *            the source object
@@ -95,14 +122,22 @@ public final class CompareObjectExtensions
 	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
 	 *             "&lt;clinit&gt;".
 	 */
-	@SuppressWarnings("rawtypes")
+	@SuppressWarnings({ "rawtypes", "unchecked" })
 	public static int compareTo(final Object sourceOjbect, final Object objectToCompare)
 		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
 	{
-		if (sourceOjbect == null || objectToCompare == null
-			|| !sourceOjbect.getClass().equals(objectToCompare.getClass()))
+		final Integer nullCheck = ComparatorExtensions.nullCheck(sourceOjbect, objectToCompare);
+		if (nullCheck != null)
 		{
-			throw new IllegalArgumentException("Object should not be null and be the same type.");
+			return nullCheck;
+		}
+		if (!sourceOjbect.getClass().equals(objectToCompare.getClass()))
+		{
+			throw new IllegalArgumentException("Object should be the same type.");
+		}
+		if (sourceOjbect instanceof Comparable && objectToCompare instanceof Comparable)
+		{
+			return ((Comparable)sourceOjbect).compareTo(objectToCompare);
 		}
 		final Map beanDescription = BeanUtils.describe(sourceOjbect);
 		beanDescription.remove("class");
@@ -111,16 +146,38 @@ public final class CompareObjectExtensions
 		int result = 0;
 		for (final Object key : beanDescription.keySet())
 		{
-			result = compareTo(sourceOjbect, objectToCompare, key.toString());
-			if (result == 0)
-			{
-				continue;
-			}
-			else
-			{
-				break;
-			}
+			result += compareTo(sourceOjbect, objectToCompare, key.toString());
+		}
+		return result;
+	}
 
+	/**
+	 * Compares the given object over the given property.
+	 *
+	 * @param sourceOjbect
+	 *            the source ojbect
+	 * @param objectToCompare
+	 *            the object to compare
+	 * @param properties
+	 *            the set of properties to compare
+	 * @return the resulted int value
+	 * @throws IllegalAccessException
+	 *             Thrown if this {@code Method} object is enforcing Java language access control
+	 *             and the underlying method is inaccessible.
+	 * @throws InvocationTargetException
+	 *             Thrown if the property accessor method throws an exception
+	 * @throws NoSuchMethodException
+	 *             Thrown if a matching method is not found or if the name is "&lt;init&gt;"or
+	 *             "&lt;clinit&gt;".
+	 */
+	public static int compareTo(final Object sourceOjbect, final Object objectToCompare,
+		final Set<String> properties)
+		throws IllegalAccessException, InvocationTargetException, NoSuchMethodException
+	{
+		int result = 0;
+		for (String property : properties)
+		{
+			result += compareTo(sourceOjbect, objectToCompare, property);
 		}
 		return result;
 	}
@@ -194,9 +251,13 @@ public final class CompareObjectExtensions
 		for (final Object key : beanDescription.keySet())
 		{
 			compareResult.put(key.toString(),
-				Integer.valueOf(compareTo(sourceOjbect, objectToCompare, key.toString())));
+				compareTo(sourceOjbect, objectToCompare, key.toString()));
 		}
 		return compareResult;
+	}
+
+	private CompareObjectExtensions()
+	{
 	}
 
 }